Fix SLO history parse error: handle object error_budget_remaining
For monitor-based SLOs the API returns error_budget_remaining as an
object {"remaining": N} instead of a plain float. Add a custom
unmarshaler that handles both forms.

+++ cmd/metrics/datadog_slos.go
@@ -104,7 +104,7 @@ func runDatadogSLOs(cmd *cobra.Command, args []string) error {
 			Type:     slo.Type,
 			Target:   target,
 			Current:  current,
-			Budget:   history.ErrorBudgetRemaining * 100,
+			Budget:   float64(history.ErrorBudgetRemaining) * 100,
 			Violated: isViolated,
 		}
 		allResults = append(allResults, r)
